requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
dj-database-url
django-cors-headers
django-crispy-forms
django-environ
django-filter
environ
gunicorn
importlib-metadata
Markdown
psycopg2
python-dateutil
six
typing_extensions
whitenoise
zipp
pymysql
djangorestframework-simplejwt